requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pytesseract==0.3.13
pillow==10.4.0
pdf2image==1.17.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
pypdf==5.0.1
gradio==4.44.0
